// app/CasterWeb/src/api/types.ts
export enum PushType {
  PUSH_TYPE_UNKNOWN = 0,
  PUSH_TYPE_NTRIP_1_0 = 1,
  PUSH_TYPE_NTRIP_2_0 = 2,
  PUSH_TYPE_TCP_CLIENT = 3,
  PUSH_TYPE_TCP_SERVER = 4,
}
// ...
export interface PushRecord {
  uid: string;
  create_time: number;
  update_time: number;
  login_mpt: string;
  type: PushType;
  target_ip: string;
  target_port: number;
  target_mpt: string;
  target_account: string;
  target_password: string;
}